🚚 Move onboarding header to onboarding screen
