+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/StepProducts.js
@@ -1,6 +1,6 @@
 import { __ } from '@wordpress/i18n';
 
-import OnboardingHeader from '../../ReusableComponents/OnboardingHeader';
+import OnboardingHeader from './Components/OnboardingHeader';
 import SelectBox from '../../ReusableComponents/SelectBox';
 import SelectBoxWrapper from '../../ReusableComponents/SelectBoxWrapper';
 import { OnboardingHooks, PRODUCT_TYPES } from '../../../data';
